backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.5.0
pymongo==3.12.3
dnspython==2.4.0
python-decouple==3.8